feat: adicionado BoardController, responsável por gerenciar movimentos e verificar  vitória
